feat(mood): moved constants in constants/ folder, removed form step4 component for now
