--- src/components/MoodForm/Step4.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Step4: React.FC = () => {
-  return <div>Step4</div>;
-};
-
-export default Step4;
