+++ src/constants/moodform.ts
@@ -1,0 +1,12 @@
+import { Emotion, Feeling } from "../types/Mood";
+
+export const emotions: Emotion[] = [
+  "anger",
+  "disgust",
+  "fear",
+  "joy",
+  "love",
+  "sadness",
+];
+
+export const feelings: Feeling[] = ["agreable", "disagreable", "neutral"];
