+++ src/utils/deleteLocalDatabase.ts
@@ -1,4 +1,4 @@
-import db from "../db";
+import db from "../constants/db";
 
 export const deleteDbFromDevice = async () => {
   try {
